pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document-anonymizer"
version = "0.1.0"
description = "General-purpose document anonymization system for scanned and digital documents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "kloia", email = "info@kloia.com"}
]
maintainers = [
    {name = "kloia", email = "info@kloia.com"}
]
keywords = ["anonymization", "document", "pdf", "masking", "gdpr", "privacy", "ocr", "pii"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Text Processing",
]
dependencies = [
    "PyMuPDF>=1.23.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "easyocr>=1.7.0",
    "httpx>=0.27.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "Faker>=24.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
preprocessing = [
    "ocr-preprocessor @ git+https://github.com/kloia/ocr-preprocessor.git@v0.1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.0",
    "types-PyYAML>=6.0",
    "types-Pillow>=10.0",
]

[project.urls]
Homepage = "https://github.com/kloia/document-anonymizer"
Repository = "https://github.com/kloia/document-anonymizer"
Documentation = "https://github.com/kloia/document-anonymizer#readme"
Changelog = "https://github.com/kloia/document-anonymizer/blob/main/CHANGELOG.md"
Issues = "https://github.com/kloia/document-anonymizer/issues"

[project.scripts]
docanon = "document_anonymizer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["document_anonymizer*"]

[tool.setuptools.package-data]
"document_anonymizer" = ["prompts/*.txt", "config/*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/document_anonymizer"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*", "*/cli.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true
